requirements.txt
gradio
pillow==7.0.0
numpy>=1.9.2
tensorflow==2.4.1

